[BSE-701] Update BodoSQL operators to return Tables instead of DataFrames (#5751)
Primary Goal: change the behavior of BodoSQL so that all operations are assumed to receive their input as a table (intead of a DataFrame), and are required to output their answer as a table (instead of a DataFrame).

Secondary Goal: Remove all uses of pd.DataFrame as a constructor; switch out with our init_dataframe utility

Ancillary Changes:

    Rewrote the BodoSQL smoke tests for convenience of unit testing (moved to 

[BSE-713] Add restrictions to getStreamingTrait #5804)

Added new BodoSQL codegen tests for several areas whose codegen was altered by this PR, but did not already have codegen tests

    approx percentile
    grouping sets
    pivot queries (moved to 

    [BSE-713] Add restrictions to getStreamingTrait #5804)

Disabled streaming for aggregations using grouping sets or where the aggregation does not have a group by (moved to

[BSE-713] Add restrictions to getStreamingTrait #5804)

Updating read_sql & similar accessors to have an extra kwarg that ensures the output is always a table (_bodo_read_as_table=True). This is needed so we can avoid thrashing that interferes with filter pushdown.

Adding a wrapper for coerce_to_array that takes in a scalar and coerces it to a replicated array (e.g. for no-groupby aggregations which output singleton arrays).

Updates to filter pushdown so that it will work directly on tables (being done in

Adding support for filter pushdown directly on tables #5813)

    Add skip markers to parquet tests with categorical data (this is a complication that will have to be resolved in a followup issue BSE-787
    Reworking MERGE INTO so that the 2 DataFrames are kept as tables until the last second, then converted into DataFrames using a specialized kernel pushdown_safe_init_df before calling the main do_delta_merge_with_target utility on them. The specialized kernel is a wrapper for init_dataframe that filter pushdown will ignore when tracking uses of a variable.
    Adds a dist argument to Table objects, just like DataFrame.
    Updates streaming snowflake/parquet tests with filter pushdown to use tables instead of DataFrames
    Updates logging utilities to display clearer error messages.
    Updating filter pushdown in typing_pass.py so it will work without pattern matching onto a call to init_dataframe, and all subsequent logic can be done directly in terms of the reader nodes instead of just the generated DF. Also allows separating the logic for the reader node vs the def node because these are done in two different places for streaming code (one of them inside the loop right before the filter, and one outside the loop).