[BSE-834] Cache lowered global variables in BodoSQL (#5890)
Adding a cache for global variables so that duplicates are not added into the BodoSQL codegen. This does not necessarily have any compilation/performance boosts, but it can the codegen files much smaller and therefore somewhat easier to read. The number of duplicate globals is currently quite high in the aftermath of the conversion from DataFrames to Tables. Any time we have an operation like sorting, union, except, intersect, grouping sets, or certain other operations that require concatenation, multiple duplicate globals are created.