[BSE-823] Fix for streaming count distinct with nullable values (#5859)
This is a specific issue with count distinct and the drop na flag. In the non-streaming path, it is always true, but in the streaming path, drop na is set to False always. After doing more investigation…

In gen_top_level_agg_func for each function:

if hasattr(func, "skipdropna"):
            skipdropna = func.skipdropna

Note that this value is passed to C++ as a single scalar. We don't seem to have the ability to control skip behavior on a per-function/ColSet basis.

func.skipdropna seems to always be true if it is set. It is set in get_agg_func if func_name in supported_agg_funcs[:-11]

Listagg is in this set, which is just wrong. I added it to a random spot in the supported_agg_funcs list, because I didn't know this func_name in supported_agg_funcs[:-11] check even existed. Additionally, I might have knocked something else out of this list that actually needs it to be set, which is REALLY wrong. Note that 'boolor_agg', 'booland_agg', 'boolxor_agg' are also in this set, which I don't know if it's intentional or not.

TLDR: For now, I've just set dropna to True. I can try and copy the existing functionality, but I'm hesitant to do so, because the exiting functionality seems a bit broken. I want to know if I should do a full re-write.

EDIT:

Pair programmed with Nick and we came to the following agreement for how to proceed:

    Clean up the varibale name for "skipdropna", and make sure it's consistent across python and C++
    Clean up set of fn’s that check for skip/dropna
    Set flag to TRUE in streaming group by WITH A VERY GOOD COMMENT explaining why it's safe