Remove BodoTZInfo from the fork (#5870)
BodoTZInfo still exists but is now included as part of a custom type
factory rather than integrated into the upstream type factory. It is
also integrated in a way that overwrites the existing
`TIMESTAMP_WITH_LOCAL_TIME_ZONE` rather than being distinct from it so
using the above type will utilize the default timezone if that is
available.

The classes and interfaces that were modified to accomodate timezone
support are now unchanged.