[BE-807] Enable printing expected costs for a query in python (#5934)
Enables an API to get the cost values directly from python. Basically just moves the RelCostWritter out of the test directory, and adds a flag to the generate_plan to enable using it instead of the default.

This is currently needed for testing TPCH plan differences on the platform, but it seems like a common thing to want to do in general.