[BSE-649]/[BSE-650]/[BSE-724] Port AVG, VAR/STD, FIRST_VALUE, LAST_VALUE to groupby.window (#5876)
Adds support for the BodoSQL window funciton `AVG` in the `groupby.window` infrastructure. Requires the usual steps to add a new function to the `groupby.window` pipeline in Python and C++, as well as a new set of aggfunc method overloads for `::mean` which will require new accumulator values `m1` (the sum as a double) and `nan_count` (how many values in the window are `NaN`. The name `m1` is used for later consistency with other moment based functions (e.g. `VARIANCE`, `SKEW` and `KURTOSIS`).

Child PRs:
[BSE-650] Adds support for the BodoSQL window functions relating to variance or standard deviation in the groupby.window infrastructure. Requires the usual steps to add a new function to the groupby.window pipeline in Python and C++, as well as a new set of aggfunc method overloads for ::var, ::var_pop, ::std and ::std_pop. This will require a new accumulator values m2 (the sum of squares of values) and heavily extend the implementation for AVG in https://github.com/Bodo-inc/Bodo/pull/5876

[BSE-724] Followup to https://github.com/Bodo-inc/Bodo/pull/5740 by adding first_value/last_value to the infrastructure developed for any_value.