Make _replace_state_definition work within the same block (#6058)
* Make _replace_state_definition work within the same block

Previously _replace_state_definition could not handle the case where the
state definition being replaced was in the same block being iterated
over to infer the type of the replacing expression. However, we already
have a mechanism to deal with this in the `_working_body` field. This
makes writing tests slightly easier.

* [run CI]

* [run CI]