[BSE-965] Support addition, subtraction and multiplication on np.matrix type (#6025)
Adding support for the operators +, -, *, @ and np.dot on matrix types. This requires dealing with array layout issues that were previously invisible. The current version of this PR modifies the Matrix type so it can use either C or F layouts, whereas before we assumed everything was C. This allows us to safely fall back on numba/numpy implementations for array operations when desired. We could try forcing everything to be C always, but this makes unboxing more complex.