[BSE-357] Convert shuffle buffers to `unique_ptr`s so that their memory can be freed earlier (#6050)
(As title)

This is particularly useful for build shuffle buffers since they will persist even after the build stage is done. We delete the join state right after the probe stage, so the probe shuffle buffer is not an issue at all.
The shuffle buffers can get large. We currently don't have a way to _reduce_ their memory usage since even `Reset` just sets the size to 0 but doesn't actually free any memory. We could add a `Free` function, but it seems cleaner to just `reset` a `unique_ptr` and let all the memory get freed automatically. 

The PR also modifies `ArrayBuildBuffer::Reset` to use `SetSize` instead of `Resize`, which is in line with our other changes.