[BSE-993] Support np.dot with float64 and int64 array arguments (#6069)
Our current support of `np.dot` requires both arguments to have the same type. This PR adds support for cases where one is an array of floats and the other is an array of integers by upcasting the integers to floats, then falling back on the regular implementation.

Child PR: [BSE-992] We already support np.linalg.norm, so this PR is about adding support for the specific case np.linalg.norm(A, axis=1) on a 2D array, which will calculate the norm of each row in A and store the results in a 1D array. Other cases do not need support at this time.