[BSE-947] Add option to move on unpin (#6105)
This PR adds a macro `MOVE_ON_UNPIN`, which when set, moves a frame to another frame in the same SizeClass as soon as it is unpinned. To build the Bodo package with this macro, run `python setup.py develop --move-on-unpin`.
We do the following in particular:
1. Find an unmapped frame in the same SizeClass. If it finds one, it moves the frame to this new unmapped frame.
2. If it cannot find an unmapped frame, it will try to find a different mapped unpinned frame in the same SizeClass. If it finds one, it will "swap" these frames, i.e. move the data around, swap the swips and update the swip owners to point to the new location.
3. If it cannot find another mapped unpinned frame, it will simply spill this frame.

I tested this with our hash and nested-loop join regression tests and it seems to work as expected. I'll open follow up PRs that introduce pipelines which use this for regular testing with this setting.
I also added prints to make sure that right things are happening, e.g. data is getting copied/swapped properly, swips are correct, etc. It was pretty hard to make (2) happen naturally, so I made it prefer to do (2) over (1) to ensure the right things were happening. Didn't get to test (3), but that's exactly the same as `--spill-on-unpin`, so I'm sure it'll work correctly.

NOTE: This is intended for correctness testing purposes only and not for production usage.