[BE-4163][BSE-1141] Use Conda-LockFile based Environment on Azure CI (#4786)
This PR refactors the "test" pipelines in Azure Nightly to use the conda-lock file to create the testing environment. This is not perfect yet, but a first step in making Azure Nightly CI more reliable and predictable. See [this comment](https://bodo.atlassian.net/browse/BSE-1141?focusedCommentId=24592) for details on some of the challenges and how we can potentially address them in follow up PRs.

Time to install all packages during Bodo test step: ~550s --> ~345s
Time to install all packages during BodoSQL test step is roughly the same (~290s; 290s -> 288s)
So this is definitely a net positive.

---------

Co-authored-by: Sahil <sahil@bodo.ai>