[BSE-1185] Minor fix for getOriginalColumnIndices (#6232)
There's a minor bug in getOriginalColumnIndices when column pruning happens within the Snowflake Rel tree. Currently, this doesn't occur for any of our tests, but it occurs fairly commonly post RelTrimmer changes.

Merging this separately because we may want this in release, and the other PR may take a while to merge. This is the specific plan where I saw this issue:

```
SnowflakeProject(BRAND_ID=[$0], CONFIRMATION_REASON=[$1], CONFIRMED_AT=[$2], ORDER_ID=[$3], RETAILER_ID=[$5])
  SnowflakeFilter(condition=[AND(SEARCH($4, Sarg['ELEVATE':VARCHAR(12), 'PRE_EXISTING':VARCHAR(12)]:VARCHAR(12)), IS NOT NULL($2))])
    SnowflakeProject(BRAND_ID=[$0], CONFIRMATION_REASON=[$1], CONFIRMED_AT=[$2], ORDER_ID=[$3], REASON=[$4], RETAILER_ID=[$5])
      SnowflakeTableScan(table=[[PRODUCTION, BRAND_RELATIONSHIPS]], Columns=[[BRAND_ID, CONFIRMATION_REASON, CONFIRMED_AT, ORDER_ID, REASON, RETAILER_ID, OUTGOING_EMAIL_TOKEN]])
```