[BSE-1217] Add MultiJoin analyzer tests (#6241)
Adds the ability to check the number of `MultiJoin` in the plan after the pass that creates the multi joins, and also the number of inputs to each `MultiJoin` in the plan. Achieves this through a series of ancillary changes:
- Creates an `AnalysisProgram` abstract class that allows its data to pass through but runs an analysis on the input that can later be retrieved.
- Creates an `AnalysisSuite` class which can contain a collection of `AnalysisProgram` objects in its companion object (currently only contains  the multi join analyzer, but this can be easily extended to include more analysis pass objects).
- Injects the multi join analyzer from the `ProgramAnalyzer` after `BodoRules.JOIN_TO_MULTI_JOIN`.  

Such analysis passes are only done when the environment variable `ENABLE_PLAN_ANALYSIS` is set. The best way to do so is to append the command line argument `-DargLine="-DENABLE_PLAN_ANALYSIS=1"` whenever calling `mvn test`