[BSE-1121] Refactor `evict` functions to use bytes instead of SizeClass index and handle an edge case in the process (#6257)
(As title)
Follow up to https://github.com/Bodo-inc/Bodo/pull/6166.
During a self-review, I noticed that we didn't properly handle the edge case where the number of bytes that need to be spilled is larger than any of the SizeClass-es. In that case, `find_size_class_idx` would return -1, which is not properly handled by the eviction functions. The best way to handle this situation is to make the evictions functions take in the number of bytes instead and then determine the SizeClass-es later. That is what this PR does. It also adds a unit test for this situation.