[BSE-1223] Correct distinctiveness information bug caused by join row ratio (#6289)
Currently, joins scale the # of distinct entries in a column by the same ratio that the # of rows increases from input to output. This makes sense when the output has fewer rows than the input, since it means that the number of distinct values could have also shrunk. However, it does not make sense to increase the # of distinct entries when the join produces more rows. Therefore, this PR now clips the ratio to scale the distinct count by to be at most 1.0. Also ensures that similar operators do not decrease the number of distinct values below 1, and that joins that could introduced nulls count as adding an extra value.

This increases the row count estimates in some locations since in some cases with back-to-back joins we were incorrectly assuming that the number of distinct values was much higher than it actually was.