[BSE-1216] Make heuristic filter pushdown irreversible (#6245)
We want to update the filter pushdown rules that are pushed down in the heuristic step to ensure they cannot be reversed. To do this, we need to ensure that our plan “locks in” our filter changes, especially for filter pushdown. This can be done by creating a variant of the SnowflakeFilter rule to support a BodoLogicalFilter directly atop a SnowflakeRel to become a SnowflakeFilter, and adding said rule to the grouping of filter pushdown rules that run before multi join.