[BSE-1344[BSE-1278] Adds tools to determine what columns need distinctiveness information (#6315)
We've been tasked with adding a number of queries + metadata to our maven CI. As a part of this task, we need to determine what columns we need approximate distinctiveness counts for. The PR adds a semi-automatic way of doing this. Running the aprox_count_distinct queries on SF and copying that information into the java files will still be manual, but this should automate the most tedious and error prone part of the process.

Specifically, This PR adds an environment variable, BODOSQL_TESTING_FIND_NEEDED_METADATA. When this environment variable is set, the testing infrastructure will track what metadata was requested, but not currently available in the system. It will then print out a number of SF queries that can be run to get the needed information.