[BSE-1240] Move the simplification rules into the HEP pass with filter pushdown (#6321)
Adds the rules for simplification and CSE into the HEP pass for filter pushdown, thus increasing the number of simplifications that are made. One query has 2 PandasFilters disappear since they are checking `IS NOT NULL` on a column we know is non-null since it was just used as the condition for an equi-join. The volcano planner was not eliminating this since removing a filter would "increase the cost" because the planner did not realize that no rows were being removed.