[BSE-1232] Simplify literals inside of ||, CONCAT, LOWER and UPPER, as well as comparisons with LEAST and GREATEST (#6328)
Allows the simplifier to do constant folding on strings inside of a call (or chain of calls) to concatenation operations. This also results in trees of `||` calls being flattened into a single call to `CONCAT`. For example:

`'%' || ' ' || T.A || '%' || '-' || 'ing'` becomes `CONCAT('% ', T.A, '%-ing')`

Child PR #6329 

Allows the simplifier to do constant folding on strings inside of a call to `LOWER` or `UPPER`. For example: `UPPER('Alphabet Soup)` becomes `'ALPHABET SOUP'`.

Child PR #6330: 

Add support for the following types of simplification patterns:
`LEAST(A, B, C) <= D` -> `A <= D OR B <= D OR C <= D`
`GREATEST(A, B, C) <= D` -> `A <= D AND B <= D AND C <= D`