[BSE-1413] Allow other Azure Nightly steps to retry on failure  (#6371)
On the most recent Nightly build, several workers failed the Setup Snowflake step. The solution is to give that step and a few similar ones the retryCountOnTaskFailure parameter so they can retry if something similar goes wrong. A value of 5 for retry attempts is used in alignment with what we already do for Maven testing.