[BSE-1360] Make many BodoSQL tests slow unless certain files are modified (#6318)
Adds a functionality `pytestmark = pytest_slow_unless_changed(...)` that takes in a regex and marks all the tests in the current file as slow unless any of the files changed in the PR are compiler passes or files that match the regex. 

This functionality was added to the BodoSQL tests for aggregation, window functions, joins, and most types of funciton coverage, as well as all of the BodoSQL array kernel files. As a result, the number of non-slow non-Snowflake tests in BodoSQL was reduced from 1632 to 443 and the number of Bodo tests from 8807 to 6985. 

Adding the markers reduced CI from roughly 115 minutes to roughly 65 minutes.

Confluence doc describing these changes & previous similar changes: https://bodo.atlassian.net/wiki/spaces/B/pages/1459257345/Conditional+Slowness+of+Testing+Suite