[BSE-1404] Fixes PandasRel conversion issue (#6368)
Fixes an issue where certain rules in Calcite were using the Default RelBuilder instead of our custom relBuilder. This resulted in plans that were invalid plans according to Calcite. The existing tests didn't catch this issue because the specific path where we used the incorrect relBuilder was never taken in any of our tests.

Specifically, this PR updates snowflake ColumnPruner and FilterLocker use the BodoLogicalBuilder. I didn't see any other rules where this was an issue.