[BSE-1419] Fix distinct row count information for SnowflakeTableScan (#6383)
Currently, SnowflakeTableScan does not account for the fact that the column numbers requested are not the same as the column numbers in the table it reads from due to column pruning. This results in us often pinging Snowflake for the incorrect columns whenever we seek distinctiveness info. This PR fixes that by adding some offset logic before doing the metadata query. This exposes that we have a huge number of columns that should have their metadata hardcoded in our testing suite but aren’t currently. This PR also adds those.