[BSE-1363]/[BSE-1379]/[BSE-1388] Expand simplification for date/time related operations (#6331)
Allows simplifying string literals into timestamp literals when casting a string to a timestamp if the string is in one of several formats:
- A date string
- A date string plus a time component (the separator between the date & time can be either a space or `T`)
- A date string plus a time component with a fraction
Any of the above can be optionally followed by a timezone specifier string, currently only allowing `+00:00`. If any other string literals are passed in that do not match these formats, the simplification attempt is abandoned.

Child PR #6345:
Allow `DATEADD` and `TIMEADD` to be simplified when they have 3 arguments and all 3 are literals, e.g. adding 72 days to a date literal or 45 minutes to a timestamp literal. Note that `DATEADD` and `TIMEADD` are aliases for one another, and both functions can accept dates, times or timestamps in their arguments. The 2-argument format (which is the same as adding/subtracting an interval) will be handled seperately.

Child PR #6350:
Expands the conditions under which `expr + interval` or `expr -  interval` can be simplified in the planner:
- Allow the interval to be a `YEAR` interval
- Allow the expr to be a timestamp, not just a date