[BSE-1393] Adds tests failing due to general join requirements to maven CI (#6379)
While investigating https://bodo.atlassian.net/browse/BSE-1393, I added several of the failing tests to the maven CI. I was unable to fix them due to needing support for general join conditions (see ticket), but we should still keep them in the CI file for if/when we do get around to handling this issue. No meta-data was included, because I didn't need it to reproduce the bugs.