[BSE-1143] Add maven unit tests for filter pushdown (attempt 2) (#6353)
Re-do of previous filter pushdown PR. Original PR can be found here: #6297. This PR was reverted due to issues with the python tests.

Fixes for said python tests were originally contained in a stacked PR here: #6354. These changes have been merged in to allow CI to pass for this PR. You can still reference the closed PR to see the changes to the original.