[BSE-1434] Partitioned Groupby: Add infrastructure for re-partitioning in the Agg case (#6430)
This PR makes the following changes:
- Adds infrastructure for splitting partitions in the AGG case. This includes implementing `SplitPartition` and adding retry loops in `UpdateGroupsAndCombine` and `FinalizeBuild` APIs, since those are the only places where a partition split can be triggered in the AGG case. 
- We have also updated the `ActivatePartition` API for the AGG case.

See the design doc for more details:
- [Processing input batches in the aggregation case](https://bodo.atlassian.net/wiki/spaces/B/pages/1449951233/Partitioned+Groupby+Spill+Support+Design+Doc#Processing-input-batches-in-the-aggregation-case)
- [Partition Split Process in the aggregation case](https://bodo.atlassian.net/wiki/spaces/B/pages/1449951233/Partitioned+Groupby+Spill+Support+Design+Doc#Partition-Split-Process-in-the-aggregation-case)
- [Handling Finalize in the aggregation case](https://bodo.atlassian.net/wiki/spaces/B/pages/1449951233/Partitioned+Groupby+Spill+Support+Design+Doc#Handling-Finalize-in-the-aggregation-case)

Note that threshold enforcement and therefore partitioning is still turned off for the AGG case. We will enable it in the next PR. The next PR is also the one that will add automated tests for this. We also need Isaac's two PRs before enabling it.
Based on some manual testing (calling `SplitPartition` manually in the middle of the loop, during `FinalizeBuild`, etc.), things seem to be working as expected. We will verify that everything works correctly when threshold enforcement is enabled.

See [design doc](https://bodo.atlassian.net/l/cp/aSSjm1eg) for more details.