[BSE-1042]/[BSE-1043] Support ARRAY_AGG on strings and ARRAY_AGG(DISTINCT) (#6188)
Adds support for the ARRAY_AGG groupby aggregation when the input array is a string array or a dictionary encoded array.

Child PR: #6200 Adds support for the ARRAY_AGG groupby aggregation with the DISTINCT keyword via a new ftype. This is accomplished by using the same setup as the regular version but by checking each element against the previous element instead of just looking for nulls. This works because DISTINCT is only allowed if the ordering clause is the same as the aggregation column. If there is no ordering clause, this implementation adds a dummy ordering (the same as the aggregaiton column) during BodoSQL codegen.