[BSE-1240] Aggregation metadata subtask: restrict available metadata (#6441)
Adds a scanning pass right before the volcano optimizer which looks for all columns where we may need their metadata in order to compute the row counts for a join. For the time being, this should not alter plans at all except for removing some of the displayed distinct counts which aren't being used. But in https://github.com/Bodo-inc/Bodo/pull/6384, this will prevent us requesting the metadata for the keys in `SELECT DISTINCT` unless the output of `SELECT DISTINCT` becomes the input to a join. Once the scanning pass is complete, any column distinctness requests that do not fall into the set generated by the scanning pass are rejected before they can ping Snowflake (or our Mockk tables).

Key forms of verification that this PR is working as intended:
- No `plan.join` files get modified
- The tests in `DistinctRequestTest.kt` still pass un-modified
- **No row counts get modified in the volcano plans**
Notably, a large number of distinctness information will disappear from the volcano plans. These should all be from columns where the distinctnss information is only being shown because the printer retrieves all metadata available from the Mockk tables, not because they are actually being requested by the planner.