[BSE-1512] Allow distinctiveness to interact with casts to timestamp (#6474)
Allows distinctiveness information to propagate through a cast to timestamp, in case it is used to infer an equi-join condition. This is the first of several additions we may make to allow function calls to propagate distinctiveness information from their input refs. By itself this does not modify the plans in any meaningful way, but it is useful setup for some reorderings that will be possible when combined with other changes.