[BSE-1659] Disable partitioning temporarily during `GroupbyState::FinalizeBuild` for partitions at max partition depth (#6560)
(As title)

We now run `Finalize` on partitions at max partition depth with partitioning (i.e. threshold enforcement) disabled.
If it wouldn't have run into a threshold enforcement error, there's no downside to doing this. If it had, then it would've tried to split the partition which would've simply raised a runtime error and halted the execution. In this situation, it's better to let it use as much memory as is available. If it fails, we're no worse off than before. However, there's a chance that this will succeed. e.g. It’s possible that this is just a hot-key case and might fit in memory after all. If the finalization of the partition succeeds, we re-enable partitioning (and threshold enforcement).
Overall, we're better off doing this. Note that this is only until we implement a proper fallback mechanism such as a sorted aggregation.