[BSE-1797] Support object_keys (#6604)
Adds support for the Snowflake JSON function OBJECT_KEYS which takes in a JSON column and returns a column where each row is an array of the string keys of that row. This is our first real JSON function in BodoSQL, so some extra testing infrastructure setup (and BdooSQL context logic) was necessary to add. Logic to toggle between map vs struct arrays was ported from this PR: #6398