[BSE-1694]/[BSE-1793]/[BSE-1799][BSE-1827] Support ARRAY_CONSTRUCT, ARRAYS_OVERLAP, ARRAY_POSITION, and EQUAL_NULL on semi-structured data (#6632)
Ensures that `EQUAL_NULL` works correctly on semi-structured data. The underlying kernel for this is a major pre-requisite for many semi-structured functions.

Child: Adds support for the Snowflake SQL function `ARRAYS_OVERLAP`. The function takes in two arrays and returns whether they have any values in common (including nulls).

Child: Adds support for the variadic Snowflake functio `ARRAY_CONSTRUCT` which can take in a variable number of arguments and packs them into an array. The snowflake version allows any types to be placed in the same array, but for now we'll only focus on types that can be easily harmonized.

Child: Adds support for the Snowflake function `ARRAY_POSITION` which takes in a value and an array and returns the zero-indexed position of the first occurrence of the value in the array (considers null a value).