[BSE-1768] Call `malloc_trim` every once in a while (#6679)
In this PR, we call `malloc_trim` after every 100MiB worth of memory has been free-d through `malloc`. `malloc_trim` forces `malloc` to return unused memory (whatever it can) back to the OS. Note that this is a Linux only feature.

See https://bodo.atlassian.net/browse/BSE-1768 for more context. TLDR:
`malloc` may not immediately return memory free-d through `free` back to the OS. It keeps the memory around for use by future 'malloc' calls. This is a standard optimization that allocators use to reduce the overall number of kernel calls. However, this can increase peak memory usage (at the OS level) substantially in certain cases, e.g. If we have a pipeline where we free a lot of small allocations (worth multiple GiBs in total) without any `malloc` calls to re-use this memory. This is what we encountered in BSE-1768. There was a pipeline where a CTB was going into a `TableBuildBuffer`. The buffers in the CTB were all allocated through `malloc`. Each buffer was quite small, but they added up to 100s of GiBs. The `TableBuildBuffer` was large enough that it was using BufferPool frames. This meant that the memory from the CTB chunks was free-d but not returned to the OS. This led to an OOM.