[BSE-1497] Support OBJECT_DELETE (#6693)
Adds support for the Snowflake SQL function OBJECT_DELETE which takes in JSON data followed by 1+ string keys and returns the JSON data without those keys. Works on struct arrays or map arrays, but on struct arrays currently requires the keys to all be passed in as string literals. Does not work on scalars properly, but that's fine for now since we don't support JSON in CASE statements in general. When we do, we'll likely need a new dict type to handle nulls in a scalar dict.