[BSE-1978] Support flatten kernel on JSON data with map arrays (#6733)
Supports FLATTEN on JSON data, so long as it is a map array. Struct arrays will need to wait for a followup since they have more complications both for typing and the runtime implementation. Note: this PR also modfies the table function definition of FLATTEN to accept type ANY in the operand type checker since the current API cannot support doing ARRAY or MAP. The return type inference rule then checks if it is an array or map type (since it needs to know where to fetch the value type from), and raises an exception if it is neither.