[BSE-1496] Support OBJECT_AGG in BodoSQL (#6682)
Adds support for the Snowflake SQL function `OBJECT_AGG` which takes in 2 column arguments and combines them into a JSON object where the first column are keys and the second column are values. This is done by producing a map array. The Snowflake API says that the second column is a variant column (so it can be heterogeneous) but for now it will be implemented as taking in arguments of any specific type (e.g. `OBJECT_AGG(string_col, int_col)` produces JSON objects with integer values). This PR only adds support for `OBJECT_AGG` under the following conditions:
- With a `GROUP BY` clause (so no table-wide or window aggregation)
- Without a `DISTINCT` clause
- On specific types of values (will need followups to address some of the more complicated array types as values, such as map arrays).