[BSE-2001] Support flatten kernel on JSON data with struct arrays (#6740)
Supports FLATTEN on JSON data that uses struct arrays, for certain combinations of value types. This is complex both on the type handling and the runtime cases because this requires combining all of the field arrays into one combined array of their interleaved values. The currently supported combinations:
- Any combination of numpy/nullable so long as numeric dtypes are compatible (makes nullable if any are nullable, otherwise numpy, and upcasts the underlying dtypes to the largest input dtype)
   - Signed int vs signed int: goes with the larger bit size, signed
   - Unsigned int vs unsigned int: goes with the larger bit size, unsigned
   - Signed int vs unsigned int: if the unsigned is the larger bit size, make keep that but make it signed. If the two have the same bit sizes or the signed is the larger bit size, go with 1 size larger than the larger size, and keep it signed.
   - Float vs float: goes with the larger bit size
   - Decimal, Int128, Date, Time, Datetime, Timedelta, Boolean: only allowed with themselves, for now
- Any combination of string / dict arrays
- Array item arrays (recursively repeats the procedure on the child arrays)

Fields that are map/struct arrays will likely be another followup for a few reasons:
- The typing rules for these are slightly more intricate than the rest.
- The runtime implementation for the struct array is a bit complicated.
- Testing these (in particular, map arrays) is a bit troublesome due to boxing/unboxing challenges in the testing infrastructure.

Note: there are no good rules to in Snowflake to compare to when it comes to how to harmonize the types since flatten is supposed to create a variant column which can contain heterogeneous values. The most important thing for maintaining Snowflake parity is avoiding casts that change the nature of the values inside the arrays (allowing integers/floats together by upcasting the integer to a float is already pushing it). This is why I, so far, have not included booleans vs other types, since boolean + int8 would require converting `false` and `true` to `0` and `1` in terms of their representations in the column, which is likely a step too far in terms of parity.