[BSE-1692] Support array_except (#6770)
Adds support for the Snowflake SQL function `ARRAY_EXCEPT` which takes in two arrays and returns all the elements from the first array that do not appear in the second. Note: if an element appears multiple times in the first array, it is only removed as many times as it appears in the second array. So `ARRAY_EXCEPT([1, 1, 2, 1, 1, 2, 3, 2, 1], [1, 2, 3, 4, 1])` would output `[1, 1, 2, 2, 1]`.

Note: since this function is one of many that uses the same testing pattern in BodoSQL, the common parametrization has been moved into a common fixture.

Child PR: #6772: Adds support for the Snowflake SQL function `ARRAY_INTERSECTION` which takes in two arrays and returns all the elements from the first array that appear in the second. Note: if an element appears multiple times in either array, it is only kept the minimum of the number of times it appears in either array. So `ARRAY_EXCEPT([1, 2, 3, 1, 2, 3], [2, 1, 1, 1, 1])` would output `[1, 1, 2]`.


Child PR #6774: Adds support for the Snowflake SQL function `ARRAY_CAT` which takes in two arrays and returns all the elements from the first array followed by the elements in the second.