[BSE-1844] Fix Union Sync Bug (#6779)
(As title)

Adds a `is_final_pipeline` flag to `union_consume_batch` and `groupby_build_consume_batch` (which `union_consume_batch` uses internally). We only call `FinalizeBuild` if `is_final_pipeline` is `True`. Updated the codegen to always pass the `is_last` from the previous operator to Union and to use the Union's `is_last` output as the pipeline terminating condition in all its pipelines (and not just the last).