[BSE-2176] Add codegen for basic flatten node with streaming (#6855)
This allows the lateral flatten kernel to be called inside of a streaming pipeline instead of requiring a combine stream exchange beforehand. A consequence of this is that one batch of 4096 inputs could output any number of rows from 0 to millions, which is not ideal if the following node in the pipeline is a streaming groupby/join. We will add a proper state to allow chunking/pausing like join does at a later date.