[BSE-2025] Add full parsing support for Snowflake's create view (#6838)
Ensure that we support parsing more of Snowflake's `CREATE VIEW` syntax. We only need to ensure that such syntax is parsed so we can use the sql string of a view definition to create the rel node subtree of the view. We don't necessarily need those features to "do" anything.

The parsing features that are to be supported:
```
CREATE [ OR REPLACE ] [ SECURE ] [ { [ { LOCAL | GLOBAL } ] TEMP | TEMPORARY | VOLATILE } ] [ RECURSIVE ] VIEW [ IF NOT EXISTS ] <name>
  [ ( <column_list> ) ]
  [ <col1> [ WITH ] MASKING POLICY <policy_name> [ USING ( <col1> , <cond_col1> , ... ) ]
           [ WITH ] TAG ( <tag_name> = '<tag_value>' [ , <tag_name> = '<tag_value>' , ... ] ) ]
  [ , <col2> [ ... ] ]
  [ [ WITH ] ROW ACCESS POLICY <policy_name> ON ( <col_name> [ , <col_name> ... ] ) ]
  [ [ WITH ] TAG ( <tag_name> = '<tag_value>' [ , <tag_name> = '<tag_value>' , ... ] ) ]
  [ COPY GRANTS ]
  [ COMMENT = '<string_literal>' ]
  AS <select_statement>
```

Where all of the keywords secure, local, global, temp, temporary, volatile, recursive, copy grants, and comment have no effect.

Where each column in the column list can be just a column name or optionally include the following:
- Type
- Nullability
- A tag (does nothing)
- A row masking policy (does nothing)

Important: does not add TAG or MASKING as non-reserved keywords because these introduce a major parser ambiguity.