[BSE-2259] Support TO_OBJECT (#6931)
Add full support for the Snowflake SQL function TO_OBJECT whose primary purpose is to unwrap variants containing an object value. The function should return its input when it is a null/map/struct scalar/vector/optional, and raise an error for every other type.

Note: the existing BodoSQL type definition does not currently handle when the input is an "unknown" type (which is a general gap, since the unknown cases often should be mapped to variant).