[BSE-2269] Add implicit casting for timestamp from parts family of functions (#6939)
Adds variant implicit casting support for the 8 functions in the TIMESTAMP_FROM_PARTS family, both in the 2-argument and the 6-8 argument version. This requires slightly changing the interface we created for adding implicit casting rules to pass in the entire operand types list because this is the first function where the implicit casting behavior considers other arguments (in this case, how many are there).